pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "little-dorrit-editor"
version = "0.1.0"
description = "Benchmark for evaluating multimodal LLMs on handwritten editorial corrections in printed text"
readme = "README.md"
requires-python = ">=3.13.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pillow>=10.1.0",
    "openai>=1.10.0",
    "datasets>=2.16.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.6.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pytest>=7.4.3",
]

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py313"
select = ["E", "F", "I", "UP", "N", "B", "A", "C4", "T20", "PT", "SIM", "RET"]
ignore = []

[tool.ruff.isort]
known-first-party = ["little_dorrit_editor"]

[tool.mypy]
python_version = "3.13"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.urls]
Repository = "https://github.com/yourusername/little-dorrit-editor"

[project.scripts]
evaluate = "little_dorrit_editor.cli:evaluate_app"
convert = "little_dorrit_editor.cli:convert_app"